+++ data/helpers/index.js
@@ -2,6 +2,7 @@ export {
   detectProjectType,
   detectProblems,
   detectUxComplaints,
+  detectCapabilities,
   detectHumanTasks,
   detectConstraints,
 } from './search.js'